requirements.txt
fastapi==0.63.0
uvicorn==0.14.0
requests==2.25.1
vk-api==11.9.1
youtube_dl==2021.5.16
redis==3.5.3
python-dotenv==0.15.0
flake8==3.9.2
pylint==2.8.3
mypy==0.902
types-redis==3.5.1
types-requests==0.1.11
black==21.6b0